refactor loading spinner tests
